Renaming the Methods for WeekTemplateRepository
After writing the WeekTemplateRepository to fit the requirements of the Controller, the respective WeekTemplateController tests were modified so that the DbContext was dependency injected and the tests were run. The result showed that instead of using the "new"ly "overwritten" methods the controller used inherited from Repository.cs(which were not written to actually work). By renaming the new methods, this problem was avoided.  The same result can be done by making methods in Repository.cs virtual.

another bug in the newly written code which was found at testing was that the method calls in the controller was not awaited which led to unassigned references. This was fixed and the WeekTemplateRepository is now fully implemented into the controller functioning as before the implementation.